move to proper folders

+++ src/Game/GameCamera.java
@@ -29,8 +29,8 @@ public class GameCamera {
 	} 
 	
 	public void centerOnEntity(Entity e) {
-		xOffset = e.getX() - handler.getWidth() / 2 + e.getWidth() / 2;
-		yOffset = e.getY() - handler.getHeight() / 2 + e.getHeight() / 2;
+		xOffset = e.getX() - (float) handler.getWidth() / 2 + (float) e.getWidth() / 2;
+		yOffset = e.getY() - (float) handler.getHeight() / 2 + (float) e.getHeight() / 2;
 		checkBlankSpace();
 	}
